cleanup pool source files and tests
